Clean AES export and move back `calculateKeyCheck` to `secret-storage.ts`
