+++ src/crypto/aes.ts
@@ -16,11 +16,8 @@ limitations under the License.
 
 import encryptAESSecretStorageItem from "../utils/encryptAESSecretStorageItem.ts";
 import decryptAESSecretStorageItem from "../utils/decryptAESSecretStorageItem.ts";
-import { AESEncryptedSecretStoragePayload } from "../@types/AESEncryptedSecretStoragePayload.ts";
 
 // Export for backwards compatibility
-export type { AESEncryptedSecretStoragePayload as IEncryptedPayload };
-// Export with new names instead of using `as` to not break react-sdk tests
-export const encryptAES = encryptAESSecretStorageItem;
-export const decryptAES = decryptAESSecretStorageItem;
-export { calculateKeyCheck } from "../calculateKeyCheck.ts";
+export type { AESEncryptedSecretStoragePayload as IEncryptedPayload } from "../@types/AESEncryptedSecretStoragePayload.ts";
+export { encryptAESSecretStorageItem as encryptAES, decryptAESSecretStorageItem as decryptAES };
+export { calculateKeyCheck } from "../secret-storage.ts";
